requirements.txt
aniso8601==9.0.1
attrs==23.1.0
blinker==1.7.0
click==8.1.7
flasgger==0.9.7.1
Flask==3.0.0
Flask-RESTful==0.3.10
flask-restx==1.3.0
Flask-Testing==0.8.1
importlib-resources==6.1.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
MarkupSafe==2.1.3
mistune==3.0.2
packaging==23.2
pluggy==1.3.0
pytest==7.4.3
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
rpds-py==0.13.2
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.9.0
Werkzeug==3.0.1
